Move input and payload types to types folder
