+++ src/graphql/schema/types/post.ts
@@ -1,6 +1,8 @@
-import { Field, ID, ObjectType } from 'type-graphql'
+import { createUnionType, Field, ID, InputType, ObjectType } from 'type-graphql'
 import { Comment } from './comment'
 import { User } from './user'
+import { UserError } from './userError'
+import { ErrorCode } from '../enums/errorCode'
 
 @ObjectType({ description: 'Represents a blog post created by a User' })
 export class Post {
@@ -22,3 +24,79 @@ export class Post {
   @Field(() => [Comment], { nullable: 'itemsAndList' })
     comments?: Comment[]
 }
+
+@InputType()
+export class CreatePostInput implements Partial<Post> {
+  @Field()
+    title: string
+
+  @Field()
+    authorId: string
+
+  @Field({ nullable: true })
+    description?: string
+}
+@ObjectType()
+export class CreatePostSuccess {
+  @Field(() => Post)
+    post: Post
+}
+@ObjectType()
+export class PostTitleTakenError implements Partial<UserError> {
+  @Field(() => ErrorCode)
+    code: ErrorCode
+
+  @Field()
+    message: string
+
+  @Field()
+    postTitleWasTaken: boolean
+}
+export const CreatePostPayload = createUnionType({
+  name: 'CreatePostPayload',
+  types: () => [CreatePostSuccess, PostTitleTakenError] as const
+})
+
+@InputType()
+export class DeletePostInput implements Partial<User> {
+  @Field(() => ID)
+    id: string
+}
+@ObjectType()
+export class DeletePostSuccess {
+  @Field(() => Post)
+    post: Post
+}
+export const DeletePostPayload = createUnionType({
+  name: 'DeletePostPayload',
+  types: () => [DeletePostSuccess, UserError] as const
+})
+
+@InputType()
+export class PublishPostInput implements Partial<User> {
+  @Field(() => ID)
+    id: string
+}
+@ObjectType()
+export class PublishPostSuccess {
+  @Field(() => Post)
+    post: Post
+}
+export const PublishPostPayload = createUnionType({
+  name: 'PublishPostPayload',
+  types: () => [PublishPostSuccess, UserError] as const
+})
+@InputType()
+export class UnpublishPostInput implements Partial<User> {
+  @Field(() => ID)
+    id: string
+}
+@ObjectType()
+export class UnpublishPostSuccess {
+  @Field(() => Post)
+    post: Post
+}
+export const UnpublishPostPayload = createUnionType({
+  name: 'UnpublishPostPayload',
+  types: () => [UnpublishPostSuccess, UserError] as const
+})
